feat(storage): implement full log_store with persistent total counter
- Add separate CTR_FILE_PATH (/spiffs/total.bin) for all-time counter
  that survives reboots and manual log clears
- Fix ring overflow: rewrite full SPIFFS dump when buffer wraps,
  append-only when space available (avoids unbounded file growth)
- Change log_store_total_count() return type int → uint32_t
- Update admin_panel.c format string %d → %lu accordingly

+++ main/admin/admin_panel.c
@@ -61,7 +61,7 @@ static void serve_attacks_json(int sock)
 {
     attack_log_t entries[50];
     int n     = log_store_get_recent(entries, 50);
-    int total = log_store_total_count();
+    uint32_t total = log_store_total_count();
 
     uint32_t seen_ips[50];
     int      unique = 0;
@@ -79,8 +79,8 @@ static void serve_attacks_json(int sock)
     static char json[8192];
     int pos = 0;
     pos += snprintf(json + pos, sizeof(json) - pos,
-        "{\"total\":%d,\"unique_ips\":%d,\"by_type\":[%d,%d,%d],\"entries\":[",
-        total, unique, by_type[0], by_type[1], by_type[2]);
+        "{\"total\":%lu,\"unique_ips\":%d,\"by_type\":[%d,%d,%d],\"entries\":[",
+        (unsigned long)total, unique, by_type[0], by_type[1], by_type[2]);
 
     for (int i = 0; i < n && pos < (int)sizeof(json) - 256; i++) {
         if (i > 0) json[pos++] = ',';
